+++ packages/api/src/router/location.ts
@@ -1,9 +1,35 @@
-import type { z } from "zod";
+import { z } from "zod";
 
 import { fetchWithTimeout } from "@homarr/common";
-import { locationSearchCityInput, locationSearchCityOutput } from "@homarr/validation/location";
 
 import { createTRPCRouter, publicProcedure } from "../trpc";
+
+const citySchema = z.object({
+  id: z.number(),
+  name: z.string(),
+  country: z.string().optional(),
+  country_code: z.string().optional(),
+  latitude: z.number(),
+  longitude: z.number(),
+  population: z.number().optional(),
+});
+
+export const locationSearchCityInput = z.object({
+  query: z.string(),
+});
+
+export const locationSearchCityOutput = z
+  .object({
+    results: z.array(citySchema),
+  })
+  .or(
+    z
+      .object({
+        generationtime_ms: z.number(),
+      })
+      .refine((data) => Object.keys(data).length === 1, { message: "Invalid response" })
+      .transform(() => ({ results: [] })), // We fallback to empty array if no results
+  );
 
 export const locationRouter = createTRPCRouter({
   searchCity: publicProcedure
